+++ core/src/raw/oio/read/into_read_from_file.rs
@@ -16,12 +16,15 @@
 // under the License.
 
 use std::cmp;
+use std::io::Read;
+use std::io::Seek;
 use std::io::SeekFrom;
 use std::pin::Pin;
 use std::task::ready;
 use std::task::Context;
 use std::task::Poll;
 
+use bytes::Bytes;
 use futures::AsyncRead;
 use futures::AsyncSeek;
 
@@ -29,10 +32,7 @@ use crate::raw::*;
 use crate::*;
 
 /// Convert given fd into [`oio::Reader`].
-pub fn into_read_from_file<R>(fd: R, start: u64, end: u64) -> FromFileReader<R>
-where
-    R: AsyncRead + AsyncSeek + Unpin + Send + Sync,
-{
+pub fn into_read_from_file<R>(fd: R, start: u64, end: u64) -> FromFileReader<R> {
     FromFileReader {
         inner: fd,
         start,
@@ -42,7 +42,7 @@ where
 }
 
 /// FdReader is a wrapper of input fd to implement [`oio::Read`].
-pub struct FromFileReader<R: AsyncRead + AsyncSeek + Unpin + Send + Sync> {
+pub struct FromFileReader<R> {
     inner: R,
 
     start: u64,
@@ -50,10 +50,7 @@ pub struct FromFileReader<R: AsyncRead + AsyncSeek + Unpin + Send + Sync> {
     offset: u64,
 }
 
-impl<R> FromFileReader<R>
-where
-    R: AsyncRead + AsyncSeek + Unpin + Send + Sync,
-{
+impl<R> FromFileReader<R> {
     pub(crate) fn current_size(&self) -> i64 {
         debug_assert!(self.offset >= self.start, "offset must in range");
         self.end as i64 - self.offset as i64
@@ -115,12 +112,72 @@ where
         }
     }
 
-    fn poll_next(&mut self, cx: &mut Context<'_>) -> Poll<Option<Result<bytes::Bytes>>> {
+    fn poll_next(&mut self, cx: &mut Context<'_>) -> Poll<Option<Result<Bytes>>> {
         let _ = cx;
 
         Poll::Ready(Some(Err(Error::new(
             ErrorKind::Unsupported,
             "output reader doesn't support seeking",
         ))))
+    }
+}
+
+impl<R> oio::BlockingRead for FromFileReader<R>
+where
+    R: Read + Seek + Send + Sync + 'static,
+{
+    fn read(&mut self, buf: &mut [u8]) -> Result<usize> {
+        if self.current_size() <= 0 {
+            return Ok(0);
+        }
+
+        let max = cmp::min(buf.len() as u64, self.current_size() as u64) as usize;
+        // TODO: we can use pread instead.
+        let n = self.inner.read(&mut buf[..max]).map_err(|err| {
+            Error::new(ErrorKind::Unexpected, "read data from FdReader")
+                .with_context("source", "FdReader")
+                .set_source(err)
+        })?;
+        self.offset += n as u64;
+        Ok(n)
+    }
+
+    /// TODO: maybe we don't need to do seek really, just call pread instead.
+    ///
+    /// We need to wait for tokio's pread support.
+    fn seek(&mut self, pos: SeekFrom) -> Result<u64> {
+        let (base, offset) = match pos {
+            SeekFrom::Start(n) => (self.start as i64, n as i64),
+            SeekFrom::End(n) => (self.end as i64, n),
+            SeekFrom::Current(n) => (self.offset as i64, n),
+        };
+
+        match base.checked_add(offset) {
+            Some(n) if n < 0 => Err(Error::new(
+                ErrorKind::InvalidInput,
+                "invalid seek to a negative or overflowing position",
+            )),
+            Some(n) => {
+                let cur = self.inner.seek(SeekFrom::Start(n as u64)).map_err(|err| {
+                    Error::new(ErrorKind::Unexpected, "seek data from FdReader")
+                        .with_context("source", "FdReader")
+                        .set_source(err)
+                })?;
+
+                self.offset = cur;
+                Ok(self.offset - self.start)
+            }
+            None => Err(Error::new(
+                ErrorKind::InvalidInput,
+                "invalid seek to a negative or overflowing position",
+            )),
+        }
+    }
+
+    fn next(&mut self) -> Option<Result<Bytes>> {
+        Some(Err(Error::new(
+            ErrorKind::Unsupported,
+            "output reader doesn't support iterating",
+        )))
     }
 }
